kvstreamer: move the package back into pkg/kv/kvclient
This commit is basically a revert of 7984db14fdadf9ad9c69107ca71b32cd7b348971
which moved `kvstreamer` package into `pkg/sql` path. The rationale was
twofold:
- a dependency on `sql/rowcontainer` - it seemed bad for a KV package to
depend on a SQL package
- SQL Queries being the owner of `kvstreamer`.

The second concern is not really a concern - it's ok for SQL Queries to
own a package living `pkg/kv` and not in `pkg/sql`.

The first concern will be addressed in the following commit.

The rationale for this commit is that the Streamer is a KV client API,
so it should live in the appropriate `pkg/kv/kvclient` path.

Release note: None
