+++ pkg/BUILD.bazel
@@ -131,6 +131,7 @@ ALL_TESTS = [
     "//pkg/keys:keys_test",
     "//pkg/kv/bulk:bulk_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord_test",
+    "//pkg/kv/kvclient/kvstreamer:kvstreamer_test",
     "//pkg/kv/kvclient/rangecache:rangecache_test",
     "//pkg/kv/kvclient/rangefeed/rangefeedbuffer:rangefeedbuffer_test",
     "//pkg/kv/kvclient/rangefeed/rangefeedcache:rangefeedcache_test",
@@ -300,7 +301,6 @@ ALL_TESTS = [
     "//pkg/sql/importer:importer_test",
     "//pkg/sql/inverted:inverted_disallowed_imports_test",
     "//pkg/sql/inverted:inverted_test",
-    "//pkg/sql/kvstreamer:kvstreamer_test",
     "//pkg/sql/lex:lex_disallowed_imports_test",
     "//pkg/sql/lex:lex_test",
     "//pkg/sql/lexbase:lexbase_test",
