+++ pkg/sql/colfetcher/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
         "//pkg/col/typeconv",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/kvclient/kvstreamer",
         "//pkg/roachpb",
         "//pkg/sql/catalog",
         "//pkg/sql/catalog/catpb",
@@ -33,7 +34,6 @@ go_library(
         "//pkg/sql/execinfra/execreleasable",
         "//pkg/sql/execinfrapb",
         "//pkg/sql/execstats",
-        "//pkg/sql/kvstreamer",
         "//pkg/sql/memsize",
         "//pkg/sql/row",
         "//pkg/sql/rowenc",
