+++ pkg/sql/rowexec/BUILD.bazel
@@ -45,6 +45,7 @@ go_library(
         "//pkg/jobs/jobspb",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/kvclient/kvstreamer",
         "//pkg/kv/kvserver/concurrency/lock",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/roachpb",
@@ -62,7 +63,6 @@ go_library(
         "//pkg/sql/execinfrapb",
         "//pkg/sql/execstats",
         "//pkg/sql/inverted",
-        "//pkg/sql/kvstreamer",
         "//pkg/sql/memsize",
         "//pkg/sql/opt/invertedexpr",
         "//pkg/sql/opt/invertedidx",
